pyproject.toml
[project]
name = "hybrid-search-demo"
version = "0.1.0"
description = "Hybrid search pipeline demo — dense + sparse + reranking + MMR over SEC filings"
requires-python = ">=3.11,<3.14"
dependencies = [
    "backoff>=2.2.0",
    "numpy>=1.24.0",
    "pandas>=2.2.0",
    "pinecone[grpc]>=7.3.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.44.0",
    "tqdm>=4.66.0",
]
